refactor runtime orchestration by platform
