+++ src/runtime/mod.rs
@@ -1,6 +1,12 @@
 pub mod ioc;
+#[cfg(target_os = "linux")]
+mod linux;
 pub mod logging;
+#[cfg(target_os = "macos")]
+mod macos;
 mod orchestration;
+#[cfg(windows)]
+mod windows;
 pub mod yara;
 
 pub use orchestration::run;
